+++ spring/umc-spring-study/src/main/java/com/example/umc/web/dto/MemberResponseDTO.java
@@ -14,7 +14,7 @@ public class MemberResponseDTO {
     @NoArgsConstructor
     @AllArgsConstructor
     public static class JoinResultDTO{
-        int memberId;
+        Long memberId;
         LocalDateTime createdAt;
     }
 }
